Update library and utilities for refactoring of 10x Genomics functions and classes.
