+++ auto_process_ngs/bcl2fastq/pipeline.py
@@ -86,12 +86,12 @@ from ..pipeliner import ListParam
 from ..pipeliner import PathExistsParam
 from ..pipeliner import PathJoinParam
 from ..pipeliner import resolve_parameter
-from ..tenx_genomics_utils import add_cellranger_args
-from ..tenx_genomics_utils import cellranger_info
-from ..tenx_genomics_utils import spaceranger_info
-from ..tenx_genomics_utils import get_bases_mask_10x_atac
-from ..tenx_genomics_utils import get_bases_mask_10x_multiome
-from ..tenx_genomics_utils import make_qc_summary_html
+from ..tenx.utils import add_cellranger_args
+from ..tenx.utils import cellranger_info
+from ..tenx.utils import spaceranger_info
+from ..tenx.utils import get_bases_mask_10x_atac
+from ..tenx.utils import get_bases_mask_10x_multiome
+from ..tenx.utils import make_qc_summary_html
 from ..utils import Location
 from ..utils import find_executables
 from .reporting import ProcessingQCReport
